requirements.txt
opencv-python
numpy
RPi.GPIO
picamera2
pycrpytodome